210805 작업
+ 피격 및 장애물 파괴시 타격감 상승 (카메라 움직임)
+ 선택한 캐릭터의 능력치가 게임에 적용되도록 수정(기존 : 기본값)
+ 시작전 아이템 구현 완료
++ 1회 피격 무시를 구매하고, 100미터 혹은 300미터 달리기를 구매한 경우,
달리기가 끝나면 1회 피격 무시가 사라지는 오류를 수정했습니다.

++ 100미터 혹은 300미터 둘 중 하나만 구매할수 있도록 수정됩니다.

++ 미 구현 되었던 랜덤 아이템 구매가 구현되었습니다.

+++ Running_Animal/Assets/Script/Character.cs
@@ -19,7 +19,7 @@ public class Character
     public int LV_JUMP_POWER;
     public float DOWN_POWER;
     public int LV_DOWN_POWER;
-    public float JUMP_COUNT;
+    public int JUMP_COUNT;
     public int LV_JUMP_COUNT;
     public float DEF;
     public int LV_DEF;
@@ -32,7 +32,7 @@ public class Character
                      float _SPEED, int _LV_SPEED,
                      float _JUMP_POWER, int _LV_JUMP_POWER,
                      float _DOWN_POWER, int _LV_DOWN_POWER,
-                     float _JUMP_COUNT, int _LV_JUMP_COUNT,
+                     int _JUMP_COUNT, int _LV_JUMP_COUNT,
                      float _DEF, int _LV_DEF,
                      int _LUK, int _LV_LUK, DataManager.Active_Skil _ACTIVE)
     {
